Contact acl issues releted to releted contacts initial commit

+++ agile-core-classes/src/com/agilecrm/core/api/calendar/EventsAPI.java
@@ -32,7 +32,9 @@ import com.agilecrm.contact.Contact;
 import com.agilecrm.contact.util.ContactUtil;
 import com.agilecrm.deals.Opportunity;
 import com.agilecrm.deals.util.OpportunityUtil;
+import com.agilecrm.projectedpojos.ContactPartial;
 import com.agilecrm.user.AgileUser;
+import com.agilecrm.user.access.exception.AccessDeniedException;
 import com.agilecrm.user.access.util.UserAccessControlUtil;
 import com.agilecrm.user.access.util.UserAccessControlUtil.CRUDOperation;
 import com.googlecode.objectify.Key;
@@ -149,28 +151,39 @@ public class EventsAPI
     {
     Event event = EventUtil.getEvent(id);
     UserAccessControlUtil.check(Event.class.getSimpleName(), event, CRUDOperation.DELETE, true);
-	try
-	{
-	    if (event != null)
-	    {
-		ActivitySave.createEventDeleteActivity(event);
-		if (event.type.toString().equalsIgnoreCase("WEB_APPOINTMENT"))
-		    GoogleCalendarUtil.deleteGoogleEvent(event);
-		  if(!(event.getDeal_ids()).isEmpty())
-	    	{
-	    		for(String oppr : event.getDeal_ids())
-	    		{
-	    			Opportunity opportuinty = OpportunityUtil.getOpportunity(Long.valueOf(oppr));
-	    			opportuinty.save();
-	    		}
-	    	}
-		event.delete();
-	    }
-	}
-	catch (Exception e)
-	{
-	    e.printStackTrace();
-	}
+    if(event != null)
+    {
+    	List<ContactPartial> contList = event.getContacts();
+    	List<String> conIds = new ArrayList<String>();
+    	for(ContactPartial con : contList)
+    	{
+    		conIds.add(String.valueOf(con.id));
+    	}
+    	List<String> modifiedConIds = UserAccessControlUtil.checkUpdateAndmodifyRelatedContacts(conIds);
+    	if(conIds != null && modifiedConIds != null && conIds.size() != modifiedConIds.size())
+    	{
+    		throw new AccessDeniedException("Sorry, you have some related contacts without update permission for contacts.");
+    	}
+    	try
+    	{
+    	    ActivitySave.createEventDeleteActivity(event);
+    		if (event.type.toString().equalsIgnoreCase("WEB_APPOINTMENT"))
+    		    GoogleCalendarUtil.deleteGoogleEvent(event);
+    		  if(!(event.getDeal_ids()).isEmpty())
+    	    	{
+    	    		for(String oppr : event.getDeal_ids())
+    	    		{
+    	    			Opportunity opportuinty = OpportunityUtil.getOpportunity(Long.valueOf(oppr));
+    	    			opportuinty.save();
+    	    		}
+    	    	}
+    		event.delete();
+    	}
+    	catch (Exception e)
+    	{
+    	    e.printStackTrace();
+    	}
+    }
     }
 
     /**
@@ -185,6 +198,12 @@ public class EventsAPI
     @Produces({ MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML })
     public Event createEvent(Event event)
     {
+    List<String> conIds = event.contacts;
+    List<String> modifiedConIds = UserAccessControlUtil.checkUpdateAndmodifyRelatedContacts(conIds);
+    if(conIds != null && modifiedConIds != null && conIds.size() != modifiedConIds.size())
+    {
+    	throw new AccessDeniedException("Sorry, you have some related contacts without update permission for contacts.");
+    }
 	event.save();
 	try
 	{
@@ -217,8 +236,27 @@ public class EventsAPI
     @Produces({ MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML })
     public Event updateEvent(Event event)
     {
-
     Event oldEvent =EventUtil.getEvent(event.id);
+    if(oldEvent != null)
+    {
+    	List<ContactPartial> contactsList = oldEvent.getContacts();
+		List<String> conIds = new ArrayList<String>();
+		for(ContactPartial cont : contactsList)
+		{
+			conIds.add(String.valueOf(cont.id));
+		}
+    	List<String> modifiedConIds = UserAccessControlUtil.checkUpdateAndmodifyRelatedContacts(conIds);
+    	if(conIds != null && modifiedConIds != null && conIds.size() != modifiedConIds.size())
+    	{
+    		throw new AccessDeniedException("Sorry, you have some related contacts without update permission for contacts.");
+    	}
+    }
+	List<String> conIds = event.contacts;
+	List<String> modifiedConIds = UserAccessControlUtil.checkUpdateAndmodifyRelatedContacts(conIds);
+	if(conIds != null && modifiedConIds != null && conIds.size() != modifiedConIds.size())
+	{
+		throw new AccessDeniedException("Sorry, you have some related contacts without update permission for contacts.");
+	}
     try {
 		if(oldEvent != null &&!(oldEvent.getDeal_ids()).isEmpty())
 		{
@@ -271,15 +309,33 @@ public class EventsAPI
     @Path("bulk")
     @POST
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
-    public void deleteEvents(@FormParam("ids") String model_ids) throws JSONException
+    @Produces({ MediaType.APPLICATION_JSON })
+    public List<String> deleteEvents(@FormParam("ids") String model_ids) throws JSONException
     {
 	JSONArray eventsJSONArray = new JSONArray(model_ids);
+	JSONArray eventsArray = new JSONArray();
+	List<String> contactIdsList = new ArrayList<String>();
     if(eventsJSONArray!=null && eventsJSONArray.length()>0){
     	try {    		
     		for(int i = 0; i < eventsJSONArray.length(); i++) {
     			
     			String eventId =  (String) eventsJSONArray.get(i);
 				Event event = EventUtil.getEvent(Long.parseLong(eventId));
+				
+				List<ContactPartial> contactsList = event.getContacts();
+				List<String> conIds = new ArrayList<String>();
+				for(ContactPartial cont : contactsList)
+				{
+					conIds.add(String.valueOf(cont.id));
+				}
+				List<String> modifiedConIds = UserAccessControlUtil.checkUpdateAndmodifyRelatedContacts(conIds);
+				
+				if(conIds == null || modifiedConIds == null || conIds.size() == modifiedConIds.size())
+				{
+					eventsArray.put(eventsJSONArray.getString(i));
+					contactIdsList.addAll(modifiedConIds);
+				}
+				
 				if(!event.getDeal_ids().isEmpty()){
 					for(String dealId : event.getDeal_ids()){
 						Opportunity oppr = OpportunityUtil.getOpportunity(Long.parseLong(dealId));
@@ -292,10 +348,11 @@ public class EventsAPI
 				e.printStackTrace();
 		}
      }
-	ActivitySave.createLogForBulkDeletes(EntityType.EVENT, eventsJSONArray,
-		String.valueOf(eventsJSONArray.length()), "");
+	ActivitySave.createLogForBulkDeletes(EntityType.EVENT, eventsArray,
+		String.valueOf(eventsArray.length()), "");
 
-	Event.dao.deleteBulkByIds(eventsJSONArray);
+	Event.dao.deleteBulkByIds(eventsArray);
+	return contactIdsList;
     }
 
     @Path("/future/list")
